Revert "feature(mapper): merged files together"
This reverts commit 5a882a1e
