+++ package.json
@@ -2,7 +2,7 @@
   "name": "two-way-mapper",
   "version": "0.8.1",
   "description": "",
-  "main": "dist/mapper.js",
+  "main": "dist/index.js",
   "files": [
     "dist",
     "src"
