+++ src/types.ts
@@ -1,0 +1,4 @@
+export type Mapper<TLeft, TRight> = {
+  map: (input: TLeft) => TRight;
+  reverse: (input: TRight) => TLeft;
+};
